requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
streamlit==1.31.0

# Google Gemini AI (Free)
google-generativeai==0.3.2

# Database
pymongo==4.6.1
motor==3.3.2

# PDF Generation
reportlab==4.0.9

# Data Models
pydantic==2.6.0
pydantic-settings==2.1.0

# HTTP & API
requests==2.31.0
httpx==0.26.0

# Environment Variables
python-dotenv==1.0.1

# Additional Utilities
python-multipart==0.0.9